Results

+++ attention/dataset_eval.py
@@ -2,19 +2,23 @@
 # Evaluate the model on the GLUE dataset
 import logging
 import os
+import random
 import shutil
 from pathlib import Path
 from typing import Any, Dict, List
 
 import datasets
+import mlflow
 import pandas as pd
 import torch
 import tqdm
 from transformers import PreTrainedModel, PreTrainedTokenizer
 
 from attention.conll import load_conllu_file, parse_to_conllu
-from attention.max_attention_weights import (heads_matching_relation,
-                                             max_attention_weights)
+from attention.max_attention_weights import (
+    heads_matching_relation,
+    max_attention_weights,
+)
 from attention.model_process import get_attention_matrix
 from attention.variability import get_relative_variability
 
@@ -22,54 +26,93 @@ logger = logging.getLogger(__file__)
 logger.setLevel(logging.DEBUG)
 
 
-# Sometimes, there's very unfrequent relations, so if there's less than this number of relations present in the analyzed data, we don't output a diagram on that
-MIN_WORDS_MATCHING_RELATION = 25
-# In decoder models, relations like this:
-# "I am human", where DEPENDANT=human and HEAD=am
-# can be modeled.
-# However, these models can't "attend to" the future.
-# So it would be impossible to obserb the following attention pattern:
-# "the green house", where DEPENDANT=green and HEAD=house
-# This is because the head comes after the dependant.
-# i.e. to obserb the relation, the head needs to come before
-# If this parameter is enabled, we also consider a "hit" the fact that HEAD attends to DEPENDANT
-# This is in line with current literature.
-ACCEPT_BIDIRECTIONAL_RELATIONS = True
 
 
-def eval_glue(model):
+def eval_glue(model, accept_bidirectional_relations):
     conll_dataset = datasets.load_dataset("glue", "cola", split="test")
-    get_matching_heads_sentence = generate_fn_get_matching_heads_sentence(model)
+    get_matching_heads_sentence = generate_fn_get_matching_heads_sentence(
+        model, accept_bidirectional_relations=accept_bidirectional_relations
+    )
     heads_matching_sentence = conll_dataset.map(
         get_matching_heads_sentence, batched=False
     )
 
     # Plot the relations
-    plot_relations(heads_matching_sentence, model=model, display=False)
+    plot_relations(heads_matching_sentence, model=model, display=False, accept_bidirectional_relations=accept_bidirectional_relations)
 
 
+def get_all_possible_conll_phrases(path_to_conll_dataset: Path):
+    # We get an input path which looks like this: '.../eu_bdt-ud-train.conllu'
+    # We want to also try to load the corresponding test and dev datasets, and concatenate them to the training dataset
+
+    all_possible_conll_phrases = load_conllu_file(path_to_conll_dataset) # This is the original dataset
+    # Now, does a test dataset exist?
+    test_dataset_path = path_to_conll_dataset.parent / (path_to_conll_dataset.name.replace("train", "test"))
+    if os.path.exists(test_dataset_path):
+        all_possible_conll_phrases += load_conllu_file(test_dataset_path)
+    else:
+        logger.warning(f"Test dataset not found at {test_dataset_path}")
+    
+    # Now, does a dev dataset exist?
+    dev_dataset_path = path_to_conll_dataset.parent / (path_to_conll_dataset.name.replace("train", "dev"))
+    if os.path.exists(dev_dataset_path):
+        all_possible_conll_phrases += load_conllu_file(dev_dataset_path)
+    else:
+        logger.warning(f"Dev dataset not found at {dev_dataset_path}")
+
+    return all_possible_conll_phrases
 def eval_ud(
-    model, tokenizer, path_to_conll_dataset, output_dir=Path(__file__).parent.parent / "results"
+    model,
+    tokenizer,
+    path_to_conll_dataset: Path,
+    output_dir=Path(__file__).parent.parent / "results",
+    trim_dataset_size=None,
+    random_seed=42,
+    **kwargs,
 ):
+    logging.info(f"Running evaluation on {path_to_conll_dataset}...")
+
+    # Check that the following arguments are defined: 'accept_bidirectional_relations', 'min_words_matching_relation'
+    if 'accept_bidirectional_relations' not in kwargs:
+        raise ValueError("The argument 'accept_bidirectional_relations' must be defined")
+    if 'min_words_matching_relation' not in kwargs:
+        raise ValueError("The argument 'min_words_matching_relation' must be defined")
+
+    # Set the random seed
+    torch.manual_seed(random_seed)
+    random.seed(random_seed)
+    mlflow.log_param("random_seed", random_seed)
+
     # Recreate the output dir - if it exists, delete it
     if os.path.exists(output_dir):
         shutil.rmtree(output_dir)
     os.makedirs(output_dir, exist_ok=True)
 
-    conll_phrases = load_conllu_file(path_to_conll_dataset)
+    conll_phrases: List[List[Dict[str, Any]]] = get_all_possible_conll_phrases(path_to_conll_dataset)
     # TODO: Definir mapeado de familias
+
+    mlflow.log_param("original_dataset_size", len(conll_phrases))
+    if trim_dataset_size:
+        # Randomly select a subset of the dataset. Choose it randomly to avoid bias
+        logger.info(
+            f"Trimming dataset from {len(conll_phrases)} to {trim_dataset_size} examples"
+        )
+        conll_phrases = random.sample(conll_phrases, trim_dataset_size)
+    mlflow.log_param("trimmed_dataset_size", len(conll_phrases))
 
     logger.info(f"About to process {len(conll_phrases)} examples...")
 
     # conll_phrases = conll_phrases[:10]
 
-    get_matching_heads_sentence = generate_fn_get_matching_heads_sentence(model, tokenizer)
+    get_matching_heads_sentence = generate_fn_get_matching_heads_sentence(
+        model, tokenizer, **kwargs
+    )
     phrases_iterator = tqdm.tqdm(conll_phrases, unit="phrase")
     heads_matching_sentence = [get_matching_heads_sentence(e) for e in phrases_iterator]
 
     # Plot the relations
     plot_relations(
-        heads_matching_sentence, model=model, display=False, output_dir=output_dir
+        heads_matching_sentence, model=model, display=False, output_dir=output_dir, **kwargs
     )
 
     uas = calculate_uas(heads_matching_sentence)
@@ -85,9 +128,17 @@ def eval_ud(
         os.makedirs(output_dir / "uas_scores", exist_ok=True)
         uas_df.to_csv(output_dir / "uas_scores" / f"uas_{relation}.csv")
 
+    # Save the UAS to MLFlow
+    # mlflow.log_table("uas", uas_df)
+    mlflow.log_artifact(output_dir / "uas_scores" / f"uas_{relation}.csv")
+
 
 # %%
-def generate_fn_get_matching_heads_sentence(model: PreTrainedModel, tokenizer: PreTrainedTokenizer):
+def generate_fn_get_matching_heads_sentence(
+    model: PreTrainedModel,
+    tokenizer: PreTrainedTokenizer,
+    **kwargs,
+):
     """
     Returns a function that takes a sentence and returns the heads matching each relation.
 
@@ -119,12 +170,14 @@ def generate_fn_get_matching_heads_sentence(model: PreTrainedModel, tokenizer: P
         dependencies_reltype = [relation for (_, _, relation) in dependencies]
 
         # Take all the words in the sentence and get the heads matching the relation
-        attention_matrix = get_attention_matrix(conll_pd=conll_pd, model=model, tokenizer=tokenizer)
+        attention_matrix = get_attention_matrix(
+            conll_pd=conll_pd, model=model, tokenizer=tokenizer
+        )
         max_weights = max_attention_weights(attention_matrix)
         heads_matching_rel = heads_matching_relation(
             conll_pd,
             max_weights,
-            accept_bidirectional_relations=ACCEPT_BIDIRECTIONAL_RELATIONS,
+            accept_bidirectional_relations=kwargs['accept_bidirectional_relations'],
         )
         # The result is a list of tuples (layer, head, head_word, dependent_word)
         # Only keep the tuples that have a dependent_word matching the relation
@@ -166,6 +219,7 @@ def plot_relations(
     model: PreTrainedModel,
     display=True,
     output_dir=(Path(__file__).parent.parent / "results"),
+    **kwargs,
 ):
     # Get all unique values for the relation
     # The relation is in the last position of the tuple 'dependencies'
@@ -208,16 +262,22 @@ def plot_relations(
                 example["dependencies_reltype"],
             ):
                 # TODO: Change this to apply to all decoder models
-                if model.config.model_type == "bloom":
+                if model.config.is_decoder:
                     # For decoder-only models, the dependant has to be posterior to the head, otherwise the attention is not possible
                     if dependant_position < head_position and (
-                        not ACCEPT_BIDIRECTIONAL_RELATIONS
+                        not kwargs['accept_bidirectional_relations']
                     ):
                         # We don't accept bidirectional relations, so it's impossible that we'll have an attention pattern DEPENDANT -> HEAD
                         logger.debug(
                             f'Skipping the dependency between {dependant_position} -> {head_position} ({example["forms"][dependant_position]} -> {example["forms"][head_position]})'
                         )
                         continue
+                else:
+                    # The model should not be an encoder-decoder. If it is, raise an error
+                    if model.config.is_encoder_decoder:
+                        raise ValueError(
+                            "The model is an encoder-decoder model. This function is not prepared to handle encoder-decoder models"
+                        )
                 if dependency == relation:
                     total_words_matching_relation += 1
             for (layer, head_position), variability in zip(
@@ -226,7 +286,7 @@ def plot_relations(
             ):
                 opacity_matrix[layer, head_position] += variability
             total_words += len(example["dependencies_reltype"])
-        if total_words_matching_relation < MIN_WORDS_MATCHING_RELATION:
+        if total_words_matching_relation < kwargs['min_words_matching_relation']:
             continue
         # opacity_matrix /= total_words  # Get the average variability for each layer and head
         # Divide the opacity matrix by the maximum value to normalize it
@@ -279,6 +339,12 @@ def plot_relations(
         # Reset the plot
         plt.clf()
 
+        mlflow.log_artifact(
+            output_dir
+            / "figures"
+            / f"heads_matching_{relation}_{model.config.model_type}.pdf"
+        )
+
         # Also store the matrix as a CSV
         matrix_df = pd.DataFrame(matrix)
         # Index column: layer
@@ -290,6 +356,12 @@ def plot_relations(
             / "csv"
             / f"heads_matching_{relation}_{model.config.model_type}.csv",
             index=True,
+        )
+
+        mlflow.log_artifact(
+            output_dir
+            / "csv"
+            / f"heads_matching_{relation}_{model.config.model_type}.csv"
         )
